+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - helper [exportToExcel](src/helpers/files/exportToExcel.ts).
   - Added prop **isSorted** for headers
+  - Added prop **isFreezeHeaders** for sheets
   - Added arg **sheetIndex** for cellCallback
 
 ---